+++ R/zzz.R
@@ -50,7 +50,7 @@ create_legal <- function() {
     paste(
       "SomaDataIO\u2122
 
-      Copyright \u00A9 2021 SomaLogic, Inc.
+      Copyright \u00A9 2022 SomaLogic Operating Co., Inc.
 
       Permission is hereby granted, free of charge, to any person obtaining
       a copy of the SomaDataIO software and associated documentation files
@@ -59,8 +59,8 @@ create_legal <- function() {
       publish, distribute, sublicense, and/or sell copies of the Software,
       and to permit persons to whom the Software is furnished to do so,
       subject to the following conditions outlined below. Further,
-      SomaDataIO and SomaLogic are trademarks owned by SomaLogic, Inc. No
-      license is hereby granted to these trademarks other than for purposes
+      SomaDataIO and SomaLogic are trademarks owned by SomaLogic Operating Co., Inc.
+      No license is hereby granted to these trademarks other than for purposes
       of identifying the origin or source of the Software.
 
       The above copyright notice and this permission notice shall be